Rename index files to component names in packages
