+++ packages/embla-carousel-auto-height/src/components/AutoHeight.ts
