+++ packages/embla-carousel-autoplay/src/components/Autoplay.ts
