+++ packages/embla-carousel-autoplay/src/index.ts
@@ -1,4 +1,4 @@
-export { AutoplayType } from './components'
+export { AutoplayType } from './components/Autoplay'
 export { AutoplayOptionsType } from './components/Options'
 
-export { default } from './components'
+export { default } from './components/Autoplay'
