+++ packages/embla-carousel-class-names/src/components/ClassNames.ts
