+++ packages/embla-carousel-class-names/src/index.ts
@@ -1,4 +1,4 @@
-export { ClassNamesType } from './components'
+export { ClassNamesType } from './components/ClassNames'
 export { ClassNamesOptionsType } from './components/Options'
 
-export { default } from './components'
+export { default } from './components/ClassNames'
