+++ packages/embla-carousel-svelte/src/components/emblaCarouselSvelte.ts
