+++ packages/embla-carousel-svelte/src/index.ts
@@ -2,6 +2,6 @@ export { EmblaOptionsType } from 'embla-carousel/components/Options'
 export { EmblaEventType } from 'embla-carousel/components/EventHandler'
 export { EmblaPluginType } from 'embla-carousel/components/Plugins'
 export { EmblaCarouselType } from 'embla-carousel/components'
-export { EmblaCarouselSvelteType } from './components'
+export { EmblaCarouselSvelteType } from './components/emblaCarouselSvelte'
 
-export { default } from './components'
+export { default } from './components/emblaCarouselSvelte'
