+++ packages/embla-carousel-vue/src/components/emblaCarouselVue.ts
