+++ packages/embla-carousel/src/components/EmblaCarousel.ts
