+++ packages/embla-carousel/src/index.ts
@@ -1,6 +1,6 @@
 export { EmblaOptionsType } from './components/Options'
 export { EmblaEventType } from './components/EventHandler'
 export { EmblaPluginType } from './components/Plugins'
-export { EmblaCarouselType } from './components'
+export { EmblaCarouselType } from './components/EmblaCarousel'
 
-export { default } from './components'
+export { default } from './components/EmblaCarousel'
